pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "floability"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "pyyaml",
  "requests"
]

[project.scripts]
floability = "floability.cli:main"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
target-version = ["py38"]